+++ package.json
@@ -10,8 +10,11 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^1.6.8",
+    "easy-peasy": "^6.0.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-intl": "^6.6.4",
     "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
