+++ src/App.jsx
@@ -1,10 +1,10 @@
 import { useState } from 'react'
 import './App.css'
 import { Route, Routes } from "react-router-dom";
-import { Home } from './Pages/Home.page';
 import React,{ ReactElement, Suspense } from "react";
 import { AppProviders } from "./providers";
-import Applayout from "./Components/applayout/Applayout.component";
+import Applayout from "./components/applayout/Applayout.component";
+import Home from './pages/Home.page';
 
 
 function App() {
